Refactor(YOLO): Renames YoloLive to YoloDarknet
YoloLive was misleading as darknet can work from a file as well as a live feed.
